Add files via upload
combined_data.rar:

Description: This file contains the combined raw data of solar radiation measurements and related meteorological and geographical features from major cities of Pakistan. The data is in a compressed RAR format and includes multiple files or data sources combined into a single archive.
Purpose: The combined data serves as the raw input for preprocessing and cleaning, allowing for the preparation of the data before training the predictive model.

cleaned_data_solar_radiation.rar:

Description: This file contains the cleaned data of solar radiation measurements in major cities of Pakistan after preprocessing of combined_data. The data has undergone cleaning processes to handle missing values, outliers, and inconsistencies. It is in a compressed RAR format for easy storage and sharing.
Purpose: The cleaned data serves as the processed input for training the predictive model and can be used for further analysis or research.

datamining_solar_radiation_cleaning (1).ipynb:

Description: This Jupyter Notebook (IPython notebook) provides the code and documentation for the preprocessing and cleaning of the solar radiation data. It includes step-by-step instructions, data manipulation, handling missing values, outlier detection, and other necessary data cleaning techniques.
Purpose: This notebook facilitates the understanding and implementation of data cleaning processes for the solar radiation dataset. It also contains additional sections that demonstrate the application of decision trees and the Apriori algorithm for practice purposes, extending beyond the primary data cleaning tasks.